+++ hovedopgave/src/assets/main.css
@@ -1,3 +1,10 @@
+/*--General--*/
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
 /*--Buttons--*/
 
 /*--Chips--*/
